requirements.txt
Flask
google-cloud-documentai
